pyproject.toml
[project]
name = "job-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docling>=2.70.0",
    "dotenv>=0.9.9",
    "exa-py>=2.1.1",
    "ipykernel>=7.1.0",
    "pydantic-ai>=1.47.0",
    "serpapi>=0.1.5",
    "pymupdf>=1.25.2",
    "rapidocr-onnxruntime>=1.4.1",
    "sqlalchemy>=2.0.41",
    "psycopg[binary]>=3.2.9",
]
